+++ App.js
@@ -30,6 +30,8 @@ import PersonalRecordsPage from "./src/Pages/PersonalRecordsPage/PersonalRecords
 import WorkoutCalendarPage from "./src/Pages/WorkoutCalendarPage/WorkoutCalendarPage";
 import SicknessPage from "./src/Pages/SicknessPage/SicknessPage";
 import SocialPostEditPage from "./src/Pages/SocialPostEditPage/SocialPostEditPage";
+import SocialPostSettingsPage from "./src/Pages/SocialPostSettingsPage/SocialPostSettingsPage";
+import ExerciseSocialPostSettingsPage from "./src/Pages/ExerciseSocialPostSettingsPage/ExerciseSocialPostSettingsPage";
 
 import { Colors } from './src/Resources/GlobalStyling/colors';
 import {
@@ -167,6 +169,8 @@ function RootNavigator() {
                 <Stack.Screen name="PersonalRecordsPage" component={PersonalRecordsPage} options={{ headerShown: false }} />
                 <Stack.Screen name="WorkoutCalendarPage" component={WorkoutCalendarPage} options={{ headerShown: false }} />
                 <Stack.Screen name="SocialPostEditPage" component={SocialPostEditPage} options={{ headerShown: false }} />
+                <Stack.Screen name="SocialPostSettingsPage" component={SocialPostSettingsPage} options={{ headerShown: false }} />
+                <Stack.Screen name="ExerciseSocialPostSettingsPage" component={ExerciseSocialPostSettingsPage} options={{ headerShown: false }} />
                 <Stack.Screen
                   name="SicknessPage"
                   component={SicknessPage}